Add consumer errorhandler class, move checkDeserializedResource, refactor schemaValidator package location
