+++ kafka-commons/src/main/java/uk/gov/homeoffice/digital/sas/kafka/consumer/errorhandler/KafkaConsumerErrorHandler.java
@@ -1,4 +1,4 @@
-package uk.gov.homeoffice.digital.sas.kafka.configuration;
+package uk.gov.homeoffice.digital.sas.kafka.consumer.errorhandler;
 
 import static uk.gov.homeoffice.digital.sas.kafka.constants.Constants.KAFKA_STOPPING_CONSUMING;
 
@@ -6,27 +6,29 @@ import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.MeterRegistry;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
 import org.springframework.kafka.config.KafkaListenerEndpointRegistry;
 import org.springframework.kafka.listener.KafkaListenerErrorHandler;
+import org.springframework.kafka.listener.ListenerExecutionFailedException;
+import org.springframework.messaging.Message;
 import org.springframework.stereotype.Component;
 import uk.gov.homeoffice.digital.sas.kafka.exceptions.KafkaConsumerException;
 
 @Component
 @Slf4j
-public class KafkaErrorHandler {
+public class KafkaConsumerErrorHandler implements KafkaListenerErrorHandler {
 
-  private final KafkaListenerEndpointRegistry kafkaListenerEndpointRegistry;
+  private MeterRegistry meterRegistry;
 
   private Counter errorCounter;
 
-  private final MeterRegistry meterRegistry;
+  private KafkaListenerEndpointRegistry kafkaListenerEndpointRegistry;
 
-  public KafkaErrorHandler(KafkaListenerEndpointRegistry kafkaListenerEndpointRegistry,
+
+  public KafkaConsumerErrorHandler(KafkaListenerEndpointRegistry kafkaListenerEndpointRegistry,
                            @Value("${actuator.kafka.failure.url}") String actuatorKafkaFailureUrl,
                            @Value("${actuator.error.type}") String actuatorErrorType,
                            @Value("${actuator.failure.description}")
-                             String actuatorKafkaFailureDescription,
+                           String actuatorKafkaFailureDescription,
                            MeterRegistry meterRegistry) {
     this.kafkaListenerEndpointRegistry = kafkaListenerEndpointRegistry;
     this.meterRegistry = meterRegistry;
@@ -34,18 +36,16 @@ public class KafkaErrorHandler {
         actuatorKafkaFailureDescription);
   }
 
-  @Bean(value = "kafkaConsumerErrorHandler")
-  public KafkaListenerErrorHandler errorHandler() {
-    return (message, exception) -> {
+  @Override
+  public Object handleError(Message<?> message, ListenerExecutionFailedException exception) {
+    // Need to throw only on chosen exception.
+    if (exception.getCause() instanceof KafkaConsumerException) {
+      errorCounter.increment();
+      log.warn(KAFKA_STOPPING_CONSUMING, exception.getCause());
+      kafkaListenerEndpointRegistry.stop();
+    }
 
-      // Need to throw only on chosen exception.
-      if (exception.getCause() instanceof KafkaConsumerException) {
-        errorCounter.increment();
-        log.warn(KAFKA_STOPPING_CONSUMING, exception.getCause());
-        kafkaListenerEndpointRegistry.stop();
-      }
-      throw exception;
-    };
+    throw exception;
   }
 
   private Counter setUpCounters(String endpointUrl, String type, String description) {
@@ -54,4 +54,6 @@ public class KafkaErrorHandler {
         .description(description)
         .register(meterRegistry);
   }
+
+
 }
